refactor: moved chartData to its own file
